@@ -1,6 +1,6 @@
 {
   "name": "tabliss",
-  "version": "2.1.0",
+  "version": "2.6.0",
   "license": "GPL-3.0",
   "private": true,
   "scripts": {
@@ -12,59 +12,52 @@
     "dev:chromium": "BUILD_TARGET=chromium webpack --watch",
     "dev:firefox": "BUILD_TARGET=firefox webpack --watch",
     "dev:web": "BUILD_TARGET=web webpack serve",
-    "test": "jest",
+    "test": "tsc -p tsconfig.test.json && jest",
     "translations": "node ./scripts/translations.js"
   },
   "dependencies": {
-    "@sentry/browser": "^6.11.0",
-    "date-fns": "^2.23.0",
-    "date-fns-tz": "^1.1.6",
+    "@sentry/browser": "^6.18.1",
+    "core-js": "^3.21.1",
+    "date-fns": "^2.28.0",
+    "date-fns-tz": "^1.3.0",
     "feather-icons": "^4.28.0",
+    "github-calendar": "^2.2.7",
     "in-browser-language": "^1.0.3",
-    "localforage": "^1.9.0",
+    "nanoid": "^3.2.0",
     "normalize.css": "^8.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-error-boundary": "^3.1.3",
-    "react-intl": "^5.20.7",
-    "react-redux": "^7.2.4",
-    "redux": "^4.1.1",
-    "redux-persist": "^6.0.0",
-    "redux-persist-webextension-storage": "^1.0.2",
-    "tlds": "^1.221.1",
-    "uuid": "^8.3.2"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-error-boundary": "^3.1.4",
+    "react-intl": "^5.25.0",
+    "tlds": "^1.228.0",
+    "webextension-polyfill": "^0.9.0"
   },
   "devDependencies": {
-    "@types/chrome": "0.0.154",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^13.11.0",
-    "@types/react": "^17.0.16",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
-    "@types/tlds": "^1.199.0",
-    "@types/uuid": "^8.3.1",
-    "clean-webpack-plugin": "^4.0.0-alpha.0",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.2.0",
-    "dotenv": "^10.0.0",
+    "@types/jest": "^27.4.0",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
+    "@types/webextension-polyfill": "^0.8.2",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^10.2.0",
+    "css-loader": "^6.5.1",
+    "dotenv": "^16.0.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.2",
-    "jest": "^27.0.6",
-    "mini-css-extract-plugin": "^2.2.0",
-    "node-sass": "^6.0.1",
-    "prettier": "2.3.2",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^27.4.7",
+    "mini-css-extract-plugin": "^2.4.6",
+    "node-sass": "^7.0.1",
+    "prettier": "2.5.1",
     "raw-loader": "^4.0.2",
     "react-intl-translations-manager": "^5.0.3",
-    "sass-loader": "^12.1.0",
-    "ts-jest": "^27.0.4",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.3.5",
+    "sass-loader": "^12.4.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.6.2",
     "typescript-react-intl": "^0.4.1",
     "url-loader": "^4.1.1",
-    "webpack": "^5.48.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-extension-reloader": "^1.1.4",
-    "workbox-webpack-plugin": "^6.2.0"
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.7.4",
+    "webpack-ext-reloader": "^1.1.9",
+    "workbox-webpack-plugin": "^6.4.2"
   }
 }
