@@ -1,17 +1,19 @@
 import { useState, useEffect } from "react";
+import { usePushError } from "../api";
 
 function areWeFullscreen() {
   return Boolean(document.fullscreenElement);
 }
 
 export function useFullscreen() {
+  const pushError = usePushError();
   const [isFullscreen, setIsFullscreen] = useState(areWeFullscreen());
 
   const toggleFullscreen = document.fullscreenEnabled
     ? () =>
         document.fullscreenElement
           ? document.exitFullscreen()
-          : document.documentElement.requestFullscreen()
+          : document.documentElement.requestFullscreen().catch(pushError)
     : false;
 
   useEffect(() => {
