@@ -1,22 +1,18 @@
 import React from "react";
-import { render } from "react-dom";
-
-import Root from "./Root";
+import { createRoot } from "react-dom/client";
 import { register as registerErrorHandler } from "./errorHandler";
 import { register as registerServiceWorker } from "./serviceWorker";
+import Root from "./views/Root";
 
 // Register error handler
-if (process.env.NODE_ENV === "production") {
+if (!DEV) {
   registerErrorHandler();
 }
 
 // Render app into root element
-render(<Root />, document.getElementById("root"));
+createRoot(document.getElementById("root")!).render(<Root />);
 
 // Register service worker on web
-if (
-  process.env.NODE_ENV === "production" &&
-  process.env.BUILD_TARGET === "web"
-) {
+if (!DEV && BUILD_TARGET === "web") {
   registerServiceWorker();
 }
