@@ -1,32 +1,22 @@
-import React, { FC, useEffect, useState } from "react";
-
+import React from "react";
+import { useObjectUrl } from "../../../hooks";
 import Backdrop from "../../../views/shared/Backdrop";
-import { Props, defaultCache } from "./types";
 import "./Image.sass";
+import { defaultCache, Props } from "./types";
 
-function pickItem(items: unknown[]) {
-  return items[Math.floor(Math.random() * items.length)];
-}
+const Image: React.FC<Props> = ({ cache = defaultCache }) => {
+  const index = React.useMemo(
+    () => Math.floor(Math.random() * cache.length),
+    [cache.length],
+  );
+  const url = useObjectUrl(cache[index]);
 
-const Image: FC<Props> = ({ cache = defaultCache }) => {
-  if (!cache.length) {
-    return <div className="Image default fullscreen" />;
-  }
-
-  const [url, setUrl] = useState<string>();
-  useEffect(() => {
-    setUrl(URL.createObjectURL(pickItem(cache)));
-
-    return () => {
-      if (url) URL.revokeObjectURL(url);
-    };
-  }, [cache]);
+  if (!url) return <div className="Image default fullscreen" />;
 
   return (
     <Backdrop
       className="Image fullscreen"
-      ready={Boolean(url)}
-      style={{ backgroundImage: url && `url(${url})` }}
+      style={{ backgroundImage: url ? `url(${url})` : undefined }}
     />
   );
 };
