@@ -1,11 +1,10 @@
-import React, { FC } from "react";
-
+import React from "react";
 import { useObjectUrls } from "../../../hooks";
 import { IconButton, RemoveIcon } from "../../../views/shared";
-import { Props, defaultCache } from "./types";
 import "./ImageSettings.sass";
+import { defaultCache, Props } from "./types";
 
-const ImageSettings: FC<Props> = ({ cache = defaultCache, setCache }) => {
+const ImageSettings: React.FC<Props> = ({ cache = defaultCache, setCache }) => {
   const urls = useObjectUrls(cache);
 
   const addImages = (files: FileList) =>
@@ -48,7 +47,7 @@ const ImageSettings: FC<Props> = ({ cache = defaultCache, setCache }) => {
         <p className="info">Large images may affect performance.</p>
       )}
 
-      <p className="info">Images do not sync between browsers.</p>
+      <p className="info">Images do not sync between devices.</p>
     </div>
   );
 };
