@@ -1,22 +1,26 @@
 import { RotatingCache } from "../../../hooks";
 import { API } from "../../types";
 
-type By = "official" | "collections" | "search";
+type By = "official" | "collections" | "search" | "topics";
 
 export interface Data {
   by: By;
   collections: string;
   featured: boolean;
+  paused?: boolean;
   search: string;
+  topics: string;
   timeout: number;
 }
 
 export interface Image {
-  data: Blob;
-  image_link: string;
-  location_title?: string;
-  user_name: string;
-  user_link: string;
+  src: string;
+  credit: {
+    imageLink: string;
+    location?: string;
+    userName: string;
+    userLink: string;
+  };
 }
 
 type Cache = RotatingCache<Image>;
@@ -27,6 +31,8 @@ export const defaultData: Data = {
   by: "official",
   collections: "",
   featured: false,
+  paused: false,
   search: "",
+  topics: "bo8jQKTaE0Y",
   timeout: 900,
 };
