@@ -12,25 +12,32 @@ import time from './time';
 import todo from './todo';
 import weather from './weather';
 import timer from './timer';
-
+import github from "./github";
+import notes from "./notes";
+import workHours from "./workHours";
+import joke from "./joke";
 
 export const widgetConfigs = [
   css,
+  github,
   greeting,
   ipInfo,
   links,
   literatureClock,
   message,
   nba,
+  notes,
   quote,
   search,
   time,
   todo,
   weather,
   timer,
+  workHours,
+  joke,
 ];
 
-if (process.env.BUILD_TARGET === "web") {
+if (BUILD_TARGET === "web") {
   widgetConfigs.push(js);
 }
 
