@@ -3,17 +3,9 @@ import React, { FC } from "react";
 import categories from "./categories";
 import { Props, defaultData } from "./types";
 
-const QuoteSettings: FC<Props> = ({ data = defaultData, setData }) => (
+const QuoteSettings: React.FC<Props> = ({ data = defaultData, setData }) => (
   <div className="QuoteSettings">
-    <label>
-      <input
-        type="radio"
-        checked={data.category === undefined}
-        onChange={() => setData({ category: undefined })}
-      />{" "}
-      All Categories
-    </label>
-
+    <h5>Daily Quotes</h5>
     {categories.map((category) => (
       <label key={category.key}>
         <input
@@ -24,7 +16,6 @@ const QuoteSettings: FC<Props> = ({ data = defaultData, setData }) => (
         {category.name}
       </label>
     ))}
-
     <p>
       Powered by{" "}
       <a
@@ -34,7 +25,18 @@ const QuoteSettings: FC<Props> = ({ data = defaultData, setData }) => (
       >
         They Said So
       </a>
-      {" and "}
+    </p>
+    <h5>Hourly Quotes</h5>
+    <label>
+      <input
+        type="radio"
+        checked={data.category === "developerexcuses"}
+        onChange={() => setData({ category: "developerexcuses" })}
+      />{" "}
+      Developer Excuses
+    </label>
+    <p>
+      Powered by{" "}
       <a
         href="http://www.developerexcuses.com/"
         target="_blank"
@@ -42,7 +44,6 @@ const QuoteSettings: FC<Props> = ({ data = defaultData, setData }) => (
       >
         Developer Excuses
       </a>
-      .
     </p>
   </div>
 );
