@@ -32,7 +32,7 @@ export default [
     name: "Students",
   },
   {
-    key: "developerexcuses",
-    name: "Developer Excuses",
+    key: "bible",
+    name: "Bible Verses",
   },
 ];
