@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 
-import { useSavedReducer, useToggle } from "../../../hooks";
+import { useKeyPress, useSavedReducer, useToggle } from "../../../hooks";
 import { DownIcon, Icon, UpIcon, ExpandIcon } from "../../../views/shared";
 import { addTodo, removeTodo, toggleTodo, updateTodo } from "./actions";
 import { reducer, State } from "./reducer";
@@ -16,6 +16,15 @@ const Todo: FC<Props> = ({ data = defaultData, setData }) => {
 
   const items = data.items.filter((item) => !item.completed || showCompleted);
   const show = !showMore ? data.show : undefined;
+
+  const keyBind = data.keyBind ?? "T";
+  useKeyPress(
+    (event: KeyboardEvent) => {
+      event.preventDefault();
+      dispatch(addTodo());
+    },
+    [keyBind.toUpperCase(), keyBind.toLowerCase()],
+  );
 
   return (
     <div className="Todo">
