@@ -1,7 +1,6 @@
 import React, { FC } from "react";
-
-import { Props, defaultData } from "./types";
 import LocationInput from "./LocationInput";
+import { defaultData, Props } from "./types";
 
 const WeatherSettings: FC<Props> = ({ data = defaultData, setData }) => (
   <div className="WeatherSettings">
@@ -11,63 +10,62 @@ const WeatherSettings: FC<Props> = ({ data = defaultData, setData }) => (
       onChange={(location) => setData({ ...data, ...location })}
     />
 
-    <label>
-      Name
-      <input
-        type="text"
-        value={data.name}
-        placeholder="Optional name"
-        onChange={(event) => setData({ ...data, name: event.target.value })}
-      />
-    </label>
+    {data.latitude && data.latitude ? (
+      <>
+        <label>
+          Name
+          <input
+            type="text"
+            value={data.name || ""}
+            placeholder="Optional name"
+            onChange={(event) =>
+              setData({ ...data, name: event.target.value || undefined })
+            }
+          />
+        </label>
 
-    <hr />
+        <hr />
 
-    <label>
-      <input
-        type="checkbox"
-        checked={data.showDetails}
-        onChange={() => setData({ ...data, showDetails: !data.showDetails })}
-      />{" "}
-      Show extended details
-    </label>
+        <label>
+          <input
+            type="checkbox"
+            checked={data.showDetails}
+            onChange={() =>
+              setData({ ...data, showDetails: !data.showDetails })
+            }
+          />{" "}
+          Show extended details
+        </label>
 
-    <label>
-      <input
-        type="radio"
-        checked={data.units === "auto"}
-        onChange={() => setData({ ...data, units: "auto" })}
-      />{" "}
-      Automatic units (based on location)
-    </label>
+        <label>
+          <input
+            type="radio"
+            checked={data.units === "si"}
+            onChange={() => setData({ ...data, units: "si" })}
+          />{" "}
+          Metric units
+        </label>
 
-    <label>
-      <input
-        type="radio"
-        checked={data.units === "si"}
-        onChange={() => setData({ ...data, units: "si" })}
-      />{" "}
-      Metric units
-    </label>
+        <label>
+          <input
+            type="radio"
+            checked={data.units === "us"}
+            onChange={() => setData({ ...data, units: "us" })}
+          />{" "}
+          Imperial units
+        </label>
 
-    <label>
-      <input
-        type="radio"
-        checked={data.units === "us"}
-        onChange={() => setData({ ...data, units: "us" })}
-      />{" "}
-      Imperial units
-    </label>
-
-    <p>
-      <a
-        href="https://darksky.net/poweredby/"
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        Powered by Dark Sky
-      </a>
-    </p>
+        <p>
+          <a
+            href="https://open-meteo.com/"
+            rel="noopener noreferrer"
+            target="_blank"
+          >
+            Weather data by Open-Meteo.com
+          </a>
+        </p>
+      </>
+    ) : null}
   </div>
 );
 
