@@ -1,50 +1,27 @@
 import { API } from "../../types";
+import { Conditions } from "./conditions";
 
 export type Coordinates = {
   latitude?: number;
   longitude?: number;
 };
 
-type ConditionIcon =
-  | "clear-day"
-  | "clear-night"
-  | "rain"
-  | "snow"
-  | "sleet"
-  | "wind"
-  | "fog"
-  | "cloudy"
-  | "partly-cloudy-day"
-  | "partly-cloudy-night";
-
-export interface Conditions {
-  alerts: {
-    title: string;
-    description: string;
-  }[];
-  apparentTemperatureHigh: number;
-  apparentTemperatureLow: number;
-  expiresAt: number;
-  humidity: number;
-  icon: ConditionIcon;
-  precipProbability: number;
-  temperatureHigh: number;
-  temperatureLow: number;
-  precipType?: number;
-  units: string;
-}
-
 export type Data = Coordinates & {
   name?: string;
   showDetails: boolean;
-  units: string;
+  units: "auto" | "si" | "us"; // `auto` has been removed, but may still be present in settings
 };
 
-export type Cache = Conditions | undefined;
+export type Cache =
+  | {
+      timestamp: number;
+      conditions: Conditions[];
+    }
+  | undefined;
 
 export type Props = API<Data, Cache>;
 
 export const defaultData: Data = {
   showDetails: false,
-  units: "auto",
+  units: "si",
 };
