@@ -1,21 +1,12 @@
-import React, { FC } from "react";
+import React from "react";
 import { FormattedMessage } from "react-intl";
-import { useDispatch } from "react-redux";
-
-import { defaultLocale } from "../../locales";
-import { useSelector } from "../../store";
-import { setLocale, setTimeZone } from "../../store/actions";
+import { db } from "../../db/state";
+import { useKey } from "../../lib/db/react";
 import TimeZoneInput from "../shared/timeZone/TimeZoneInput";
 
-const System: FC = () => {
-  const locale = useSelector((state) => state.data.locale || defaultLocale);
-  const timeZone = useSelector((state) => state.data.timeZone || "");
-
-  const dispatch = useDispatch();
-  const handleSetLocale = (event: React.ChangeEvent<HTMLSelectElement>) =>
-    dispatch(setLocale(event.target.value));
-  const handleSetTimeZone = (timeZone?: string) =>
-    dispatch(setTimeZone(timeZone));
+const System: React.FC = () => {
+  const [locale, setLocale] = useKey(db, "locale");
+  const [timeZone, setTimeZone] = useKey(db, "timeZone");
 
   return (
     <div>
@@ -38,7 +29,10 @@ const System: FC = () => {
         }}
       >
         <span>Language</span>
-        <select value={locale} onChange={handleSetLocale}>
+        <select
+          value={locale}
+          onChange={(event) => setLocale(event.target.value)}
+        >
           <option value="ca-ES" title="Catalan">
             Català
           </option>
@@ -54,6 +48,9 @@ const System: FC = () => {
           <option value="en-AU" title="English (Australian)">
             English (AU)
           </option>
+          <option value="en-CA" title="English (Canadian)">
+            English (CA)
+          </option>
           <option value="en-GB" title="English (British)">
             English (GB)
           </option>
@@ -63,14 +60,26 @@ const System: FC = () => {
           <option value="es" title="Spanish">
             Español
           </option>
-          <option value="fa" title="French">
+          <option value="fa" title="Persian">
             پارسی
           </option>
           <option value="fr" title="French">
             Français
           </option>
+          <option value="ga" title="Gaeilge">
+            Gaeilge
+          </option>
           <option value="gd" title="Scottish Gaelic">
             Gàidhlig
+          </option>
+          <option value="gl" title="Galician">
+            Galego
+          </option>
+          <option value="gu" title="Gujarati">
+            ગુજરાતી
+          </option>
+          <option value="hi" title="Hindi">
+            हिन्दी
           </option>
           <option value="hu" title="Hungarian">
             Magyar
@@ -92,6 +101,9 @@ const System: FC = () => {
           </option>
           <option value="lt" title="Lithuanian">
             Lietuvių k.
+          </option>
+          <option value="ne" title="Nepali">
+            Nepali
           </option>
           <option value="nl" title="Dutch">
             Nederlands
@@ -117,6 +129,9 @@ const System: FC = () => {
           <option value="sk" title="Slovak">
             Slovenčina
           </option>
+          <option value="sr" title="Serbian">
+            Српски
+          </option>
           <option value="fi" title="Finnish">
             Suomi
           </option>
@@ -141,12 +156,6 @@ const System: FC = () => {
           <option value="zh-TW" title="Traditional Chinese (Taiwan)">
             中文（台灣）
           </option>
-          <option value="hi" title="Hindi">
-            हिन्दी
-          </option>
-          <option value="gu" title="Gujarati">
-            ગુજરાતી
-          </option>
           <option value="uk" title="Ukrainian">
             Українська
           </option>
@@ -164,7 +173,7 @@ const System: FC = () => {
         }}
       >
         Time Zone
-        <TimeZoneInput timeZone={timeZone} onChange={handleSetTimeZone} />
+        <TimeZoneInput timeZone={timeZone} onChange={setTimeZone} />
       </label>
     </div>
   );
