@@ -1,12 +1,10 @@
-import React, { FC } from "react";
-
+import React from "react";
 import "./Logo.css";
+import tablissLogo from "./tabliss.svg";
 
-const tablissLogo = require("./tabliss.svg");
-
-const Logo: FC = () => (
+const Logo: React.FC = () => (
   <h1 className="Logo">
-    <i dangerouslySetInnerHTML={{ __html: tablissLogo.default }} />
+    <i dangerouslySetInnerHTML={{ __html: tablissLogo }} />
   </h1>
 );
 
