@@ -1,6 +1,5 @@
-import React, { FC } from "react";
+import React from "react";
 import { withErrorBoundary } from "react-error-boundary";
-
 import { capture as captureException } from "../../errorHandler";
 import { useApi } from "../../hooks";
 import { API } from "../../plugins";
@@ -11,7 +10,7 @@ type Props = {
   component: React.ComponentType<API<any, any>>;
 };
 
-const Plugin: FC<Props> = ({ id, component: Component }) => {
+const Plugin: React.FC<Props> = ({ id, component: Component }) => {
   // Create plugin API
   const api = useApi(id);
 
