@@ -1,11 +1,17 @@
-import React, { FC } from "react";
-
+import React from "react";
 import "./Modal.css";
 
-const Modal: FC = ({ children }) => {
+type Props = {
+  children: React.ReactNode;
+  onClose: () => void;
+};
+
+const Modal: React.FC<Props> = ({ children, onClose }) => {
   return (
-    <div className="Modal-container">
-      <div className="Modal">{children}</div>
+    <div className="Modal-container" onClick={onClose}>
+      <div className="Modal" onClick={(event) => event.stopPropagation()}>
+        {children}
+      </div>
     </div>
   );
 };
