@@ -1,50 +1,63 @@
-import { utcToZonedTime } from "date-fns-tz";
-import React, { FC, useEffect, useState } from "react";
-
-import timeZones from "./timeZones";
-
-let cachedSupportedTimeZones: string[] | undefined;
+import { getTimezoneOffset } from "date-fns-tz";
+import React from "react";
+import zones from "./zones.json";
 
 type Props = {
-  timeZone?: string;
-  onChange: (timeZone?: string) => void;
+  timeZone: string | null;
+  onChange: (timeZone: string | null) => void;
 };
 
-const TimeZoneInput: FC<Props> = ({ timeZone, onChange }) => {
-  const [supportedTimeZones, setSupportedTimeZones] = useState(
-    cachedSupportedTimeZones,
-  );
+type ZoneOption = {
+  id: string;
+  name: string;
+  offset: number;
+};
 
-  useEffect(() => {
-    if (!supportedTimeZones) {
-      // Generate supported timezones
-      cachedSupportedTimeZones = timeZones.filter((timeZone) => {
+let cachedZoneOptions: ZoneOption[] | null = null;
+
+const TimeZoneInput: React.FC<Props> = ({ timeZone, onChange }) => {
+  const [zoneOptions, setZoneOptions] = React.useState(cachedZoneOptions);
+
+  React.useEffect(() => {
+    if (zoneOptions !== null) return;
+
+    const date = new Date();
+    cachedZoneOptions = zones
+      .flatMap((zone) => {
         try {
-          utcToZonedTime(new Date(), timeZone);
+          const offset = getTimezoneOffset(zone, date) / 3_600_000;
+          const offsetFormatted = (offset >= 0 ? "+" : "-") + Math.abs(offset);
+          return {
+            id: zone,
+            name: `(UTC${offsetFormatted}) ${zone.replaceAll("_", " ")}`,
+            offset,
+          };
         } catch {
-          return false;
+          // This time zone not supported in this browser
+          return [];
         }
-        return true;
+      })
+      .sort((a, b) => {
+        const delta = a.offset - b.offset;
+        return delta === 0 ? a.name.localeCompare(b.name) : delta;
       });
-
-      setSupportedTimeZones(cachedSupportedTimeZones);
-    }
+    setZoneOptions(cachedZoneOptions);
   }, []);
 
   return (
     <select
-      value={timeZone}
-      onChange={(event) => onChange(event.target.value || undefined)}
+      value={timeZone || ""}
+      onChange={(event) => onChange(event.target.value || null)}
     >
       <option value="">Automatic</option>
-
-      {supportedTimeZones &&
-        supportedTimeZones.map((timeZone) => (
-          <option key={timeZone}>{timeZone}</option>
-        ))}
-
-      {!supportedTimeZones && (
-        <option disabled>Calculating supported time zones...</option>
+      {zoneOptions ? (
+        zoneOptions.map((option) => (
+          <option key={option.id} value={option.id}>
+            {option.name}
+          </option>
+        ))
+      ) : (
+        <option disabled>Loading...</option>
       )}
     </select>
   );
